use sane date formatting for agenda, show relative time
Use sane ISO-like formatting for the agenda times instead of the
wasteful abomination that spent a dozen or so characters to spell out
the name of the day (I'm pretty sure I can figure out that "Wed" in fact
means "Wednesday") and week's number (which I find completely useless).
Of course this was accompanied by the number of the year (another five
characters of wasted space). Now it is all compacted into ~MM-DD~.

Also added relative hour time for all the tasks shown in agenda
formatting.